Removed pending handler from Screen API.
